backend/requirements.txt
Django>=5.0.0,<6.0.0
djangorestframework>=3.15.0,<4.0.0
django-cors-headers>=4.3.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
PyJWT>=2.8.0,<3.0.0
google-generativeai>=0.3.0 